p4i integration Integrate 241719 from 4.2 to main: Fixes:    compile RevBy:    Martin Smith Details:  don't assume qreal == float
[git-p4: depot-paths = "//depot/qt/main/": change = 241720]

+++ src/gui/painting/qbezier.cpp
@@ -290,7 +290,7 @@ static ShiftResult shift(const QBezier *orig, QBezier *shifted, qreal offset, qr
         qreal r = 1.0 + prev_normal.x() * next_normal.x()
                       + prev_normal.y() * next_normal.y();
 
-        if (qFuzzyCompare(r, 0.0)) {
+        if (qFuzzyCompare(r, (qreal)0.0)) {
             points_shifted[i] = points[i] + offset * prev_normal;
         } else {
             qreal k = offset / r;
